move some methods to field_resolver module
